Implemented Databse and Providers for all models.
Added implementation of Database for all models to provide access to the Firestore Database.

Added updated providers to reflect the access to the database update.

Also added all data providers, ConnectUni loading and error widgets.

Ignore the files that are not .dart, using them as a temporary way of saving what the old providers did to reimplement them in the next commit.